--- src/app/account/signin/page.tsx
@@ -1,107 +1,0 @@
-"use client";
-
-import Image from "next/image";
-import googleImg from "@public/images/accounts/google.png";
-import Base_Nav from "@components/base/_base_nav";
-import { useRouter } from "next/navigation";
-
-export default function Home() {
-  const router = useRouter();
-
-  return (
-    <section className="mb-0">
-      <Base_Nav />
-      <main className="w-full h-full max-w-[68em] m-auto max-md:max-w-full">
-        <section className="accounts-signin  w-full max-sm:sm:p-0 mb-2">
-          <div className="accounts-signin-container p-[1em] max-md:md:p-0">
-            <div className="max-w-[33em] bg-white m-auto box-shadow-1 rounded-md p-[1em] md:max-md:w-[90%]">
-              <span className="title-gradient text-4xl text-center w-full block py-3 font-bold">
-                Welcome Back
-              </span>
-              <div className="form-group py-3 flex flex-col gap-3">
-                <div className="accounts-form-group">
-                  <label htmlFor="username">Username</label>
-                  <div className="flex">
-                    <span className="inline-flex items-center px-3 text-sm text-gray-900 bg-gray-200 border border-r-0 border-gray-300 rounded-l-md dark:bg-gray-600 dark:text-gray-400 dark:border-gray-600">
-                      <svg
-                        className="w-5 h-5 text-gray-500 dark:text-gray-400"
-                        aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg"
-                        fill="currentColor"
-                        width="16px"
-                        viewBox="0 0 20 20"
-                      >
-                        <path d="M10 0a10 10 0 1 0 10 10A10.011 10.011 0 0 0 10 0Zm0 5a3 3 0 1 1 0 6 3 3 0 0 1 0-6Zm0 13a8.949 8.949 0 0 1-4.951-1.488A3.987 3.987 0 0 1 9 13h2a3.987 3.987 0 0 1 3.951 3.512A8.949 8.949 0 0 1 10 18Z" />
-                      </svg>
-                    </span>
-                    <input
-                      autoComplete="off"
-                      type="text"
-                      id="username"
-                      required
-                      className="rounded-none outline-none rounded-r-lg bg-gray-50 border text-gray-900 focus:ring-foreground-accent-color-2 focus:border-foreground-accent-color-2 block flex-1 min-w-0 w-full text-sm border-gray-300 p-2.5  dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-foreground-accent-color-2 dark:focus:border-foreground-accent-color-2"
-                      placeholder="johnwick"
-                    />
-                  </div>
-                </div>
-
-                <div className="accounts-form-group">
-                  <label htmlFor="password">Password</label>
-                  <div className="flex">
-                    <span className="inline-flex items-center px-3 text-sm text-gray-900 bg-gray-200 border border-r-0 border-gray-300 rounded-l-md dark:bg-gray-600 dark:text-gray-400 dark:border-gray-600">
-                      <svg
-                        className="w-5 h-5 text-gray-500 dark:text-gray-400"
-                        xmlns="http://www.w3.org/2000/svg"
-                        viewBox="0 0 24 24"
-                        fill="none"
-                      >
-                        <path
-                          fillRule="evenodd"
-                          clipRule="evenodd"
-                          d="M3.17157 5.17157C2 6.34315 2 8.22876 2 12C2 15.7712 2 17.6569 3.17157 18.8284C4.34315 20 6.22876 20 10 20H14C17.7712 20 19.6569 20 20.8284 18.8284C22 17.6569 22 15.7712 22 12C22 8.22876 22 6.34315 20.8284 5.17157C19.6569 4 17.7712 4 14 4H10C6.22876 4 4.34315 4 3.17157 5.17157ZM12.7502 10C12.7502 9.58579 12.4144 9.25 12.0002 9.25C11.586 9.25 11.2502 9.58579 11.2502 10V10.7012L10.6428 10.3505C10.2841 10.1434 9.8254 10.2663 9.61829 10.625C9.41119 10.9837 9.53409 11.4424 9.89281 11.6495L10.4997 11.9999L9.89258 12.3505C9.53386 12.5576 9.41095 13.0163 9.61806 13.375C9.82517 13.7337 10.2839 13.8566 10.6426 13.6495L11.2502 13.2987V14C11.2502 14.4142 11.586 14.75 12.0002 14.75C12.4144 14.75 12.7502 14.4142 12.7502 14V13.2993L13.3569 13.6495C13.7156 13.8566 14.1743 13.7337 14.3814 13.375C14.5885 13.0163 14.4656 12.5576 14.1069 12.3505L13.4997 11.9999L14.1067 11.6495C14.4654 11.4424 14.5883 10.9837 14.3812 10.625C14.1741 10.2663 13.7154 10.1434 13.3567 10.3505L12.7502 10.7006V10ZM6.73266 9.25C7.14687 9.25 7.48266 9.58579 7.48266 10V10.7006L8.0891 10.3505C8.44782 10.1434 8.90651 10.2663 9.11362 10.625C9.32073 10.9837 9.19782 11.4424 8.8391 11.6495L8.23217 11.9999L8.83934 12.3505C9.19806 12.5576 9.32096 13.0163 9.11386 13.375C8.90675 13.7337 8.44806 13.8566 8.08934 13.6495L7.48266 13.2993V14C7.48266 14.4142 7.14687 14.75 6.73266 14.75C6.31844 14.75 5.98266 14.4142 5.98266 14V13.2987L5.375 13.6495C5.01628 13.8566 4.55759 13.7337 4.35048 13.375C4.14337 13.0163 4.26628 12.5576 4.625 12.3505L5.23217 11.9999L4.62523 11.6495C4.26652 11.4424 4.14361 10.9837 4.35072 10.625C4.55782 10.2663 5.01652 10.1434 5.37523 10.3505L5.98266 10.7012V10C5.98266 9.58579 6.31844 9.25 6.73266 9.25ZM18.0181 10C18.0181 9.58579 17.6823 9.25 17.2681 9.25C16.8539 9.25 16.5181 9.58579 16.5181 10V10.7012L15.9106 10.3505C15.5519 10.1434 15.0932 10.2663 14.8861 10.625C14.679 10.9837 14.8019 11.4424 15.1606 11.6495L15.7676 11.9999L15.1604 12.3505C14.8017 12.5576 14.6788 13.0163 14.8859 13.375C15.093 13.7337 15.5517 13.8566 15.9104 13.6495L16.5181 13.2987V14C16.5181 14.4142 16.8539 14.75 17.2681 14.75C17.6823 14.75 18.0181 14.4142 18.0181 14V13.2993L18.6247 13.6495C18.9835 13.8566 19.4422 13.7337 19.6493 13.375C19.8564 13.0163 19.7335 12.5576 19.3747 12.3505L18.7676 11.9999L19.3745 11.6495C19.7332 11.4424 19.8561 10.9837 19.649 10.625C19.4419 10.2663 18.9832 10.1434 18.6245 10.3505L18.0181 10.7006V10Z"
-                          fill="grey"
-                        />
-                      </svg>
-                    </span>
-                    <input
-                      autoComplete="off"
-                      type="password"
-                      required
-                      className="rounded-none outline-none  rounded-r-lg bg-gray-50 border text-gray-900 focus:ring-foreground-accent-color-2 focus:border-foreground-accent-color-2 block flex-1 min-w-0 w-full text-sm border-gray-300 p-2.5  dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-foreground-accent-color-2 dark:focus:border-foreground-accent-color-2"
-                      id="password"
-                      placeholder="********"
-                    />
-                  </div>
-                </div>
-
-                <button
-                  type="submit"
-                  className="mt-3 border-none box-shadow-1 outline-none bg-foreground-accent-color-3 w-full h-[2.5em] text-white hover:bg-foreground-accent-color-2/90 font-bold rounded-md"
-                >
-                  Login
-                </button>
-
-                <a
-                  onClick={() => router.push("/account/signup")}
-                  className="font-semibold underline text-foreground-accent-color-4 text-right hover:cursor-pointer hover:text-foreground-accent-color-3"
-                >
-                  New User?
-                </a>
-
-                <span className="m-auto text-gray-600/60"> --OR-- </span>
-
-                <div className="accounts-signin-google border-2 p-2 w-[3em] h-[3em] rounded-full cursor-pointer m-auto hover:bg-foreground-accent-color-2/50">
-                  <Image
-                    src={googleImg}
-                    alt="google-auth"
-                    className="w-full h-full rounded-full"
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-      </main>
-    </section>
-  );
-}
