--- src/app/account/signup/constant/signupErrorConstant.ts
@@ -1,10 +1,0 @@
-export const usernameLengthErrorMessage = "Username must be greater than 8 characters.";
-export const usernameWhitespaceErrorMessage = "Username cannot contain whitespace.";
-export const usernameSpecialCharErrorMessage =
-  "Username cannot contain special characters.";
-export const usernameNumberErrorMessage = "Username can have at least 3 numbers.";
-
-export const passwordLengthErrorMessage = "Password must be greater than 8 characters.";
-export const passwordSpecialCharErrorMessage =  "Password must contain at least one special character.";
-export const passwordCapitalLetterErrorMessage = "Password must contain at least one capital letter.";
-export const passwordNumberErrorMessage = "Password must contain at least one number.";
