--- src/app/account/signup/enum/singup.enum.ts
@@ -1,10 +1,0 @@
-export enum ESignupConstant {
-  INPUT_EMAIL = "email",
-  INPUT_USERNAME = "username",
-  INPUT_PASSWORD = "password",
-}
-
-export enum ESignupErrorTypeConstant {
-  USERNAME_ERROR_TYPE = 'username-error-type',
-  PASSWORD_ERROR_TYPE = 'password-error-type',
-}
