--- src/app/account/signup/interface/signup.interfaces.ts
@@ -1,3 +1,0 @@
-export default interface IValidateCallback {
-  (id: string, onClick?: boolean): void;
-}
