--- src/constants/auth-constants/auth.constant.interface.ts
@@ -1,15 +1,0 @@
-export interface UserType {
-  email: string | null;
-  uid: string | null;
-}
-
-export interface LoginType {
-  email: string | null;
-  password: string | null;
-}
-
-export interface RegistrationType {
-  email: string | null;
-  password: string | null;
-  username: string | null;
-}
