--- src/hooks/context/protected-route.tsx
@@ -1,20 +1,0 @@
-'use client';
-
-import { useAuth } from "src/hooks/context/AuthContext";
-import { useRouter } from "next/navigation";
-import React, { useEffect } from "react";
-
-const ProtectedRoute = ({ children }: { children: React.ReactNode }) => {
-  const router = useRouter();
-  const { user } = useAuth();
-
-  useEffect(() => {
-    if (!user.uid) {
-      router.push("/account/signin");
-    }
-  }, [router, user]);
-
-  return <div>{user ? children : null}</div>;
-};
-
-export default ProtectedRoute;
